Hopefully fixed the bug so that the mismatch pictures for SSD will work. We had an overflow which meant that really bad matches were counting as the closest, since their distance was so large it overflowed to negative.

+++ image.py
@@ -83,7 +83,7 @@ def scramble(img, type2=True):
 	for i in xrange(num_squares_H):
 		for j in xrange(num_squares_W):
 			pixels = img.reduced[i*P:(i+1)*P, j*P:(j+1)*P, :]
-			pixels = pixels.astype(np.int16)
+			pixels = pixels.astype(np.int32)
 			rot_idx = 0
 			if type2:
 				rot_idx = random.randint(0, 3)
